Feature/hibernate for models and repositories

+++ src/main/java/ru/rasim/controllers/BooksController.java
@@ -29,7 +29,7 @@ public class BooksController {
     }
 
     @GetMapping("/{id}")
-    public String show(@PathVariable("id") Integer id, Model model) {
+    public String show(@PathVariable("id") Long id, Model model) {
         model.addAttribute("book", booksRepository.show(id));
         return "book/show";
     }
@@ -50,14 +50,14 @@ public class BooksController {
     }
 
     @GetMapping("/{id}/edit")
-    public String edit(@PathVariable("id") Integer id, Model model) {
+    public String edit(@PathVariable("id") Long id, Model model) {
         model.addAttribute("book", booksRepository.show(id));
         return "book/edit";
     }
 
     @PatchMapping("/{id}")
     public String update(@ModelAttribute("book") @Valid Book book, BindingResult bindingResult,
-                         @PathVariable("id") int id) {
+                         @PathVariable("id") Long id) {
         if (bindingResult.hasErrors())
             return "book/edit";
 
@@ -66,7 +66,7 @@ public class BooksController {
     }
 
     @DeleteMapping("/{id}")
-    public String delete(@PathVariable("id") int id) {
+    public String delete(@PathVariable("id") Long id) {
         booksRepository.delete(id);
         return "redirect:/books";
     }
