+++ src/main/java/ru/rasim/repositories/BookingsRepository.java
@@ -5,11 +5,11 @@ import ru.rasim.models.Booking;
 import java.util.List;
 
 public interface BookingsRepository extends CrudRepository<Booking> {
-    boolean finish(Integer id);
+    boolean finish(Long id);
 
     List<Booking> showAllActive();
 
     List<Booking> showAllFinished();
 
-    List<Booking> showByPersonId(Integer id);
+    List<Booking> showByPersonId(Long id);
 }
